Undo refactoring of link.rs
